+++ client/src/routes.tsx
@@ -3,7 +3,7 @@ import { Route } from "react-router";
 import { BrowserRouter, Switch } from "react-router-dom";
 
 import Authentication from "./views/authentication";
-import MappingExplorer from "./views/mappingExplorer";
+import Mapping from "./views/mapping";
 import MimicView from "./views/mimic";
 import NewSource from "./views/newSource";
 import Sources from "./views/sources";
@@ -13,7 +13,7 @@ const Routes = () => (
     <Switch>
       <Route exact path="/" component={Authentication} />
       <Route path="/signin" component={Authentication} />
-      <Route path="/mapping" component={MappingExplorer} />
+      <Route path="/mapping" component={Mapping} />
       <Route path="/mimic" component={MimicView} />
       <Route path="/newSource" component={NewSource} />
       <Route path="/sources" component={Sources} />
