+++ client/src/views/mapping/components/inputColumn.tsx
