+++ client/src/views/mapping/components/inputColumns.tsx
