+++ client/src/views/mapping/components/join.tsx
