+++ client/src/views/mapping/components/joinColumns.tsx
